refactoring of topologygraph
