+++ org.eclipse.winery.topologygraph/pom.xml
@@ -33,7 +33,7 @@
         <dependency>
             <groupId>org.jgrapht</groupId>
             <artifactId>jgrapht-core</artifactId>
-            <version>1.0.1</version>
+            <version>${org.jgrapht.jgrapht-core.version}</version>
         </dependency>
         <dependency>
             <groupId>org.eclipse.winery</groupId>
@@ -41,6 +41,5 @@
             <version>2.0.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
-
 
 </project>
