+++ org.eclipse.winery.topologygraph/src/main/java/org/eclipse/winery/topologygraph/matching/ToscaComparatorProxy.java
@@ -13,21 +13,21 @@
  ********************************************************************************/
 package org.eclipse.winery.topologygraph.matching;
 
-import org.eclipse.winery.topologygraph.matching.model.ToscaEdge;
-import org.eclipse.winery.topologygraph.matching.model.ToscaNode;
+import org.eclipse.winery.topologygraph.model.ToscaEdge;
+import org.eclipse.winery.topologygraph.model.ToscaNode;
 
 public class ToscaComparatorProxy {
-	private final IToscaMatcher matcher;
+    private final IToscaMatcher matcher;
 
-	public ToscaComparatorProxy(IToscaMatcher matcher) {
-		this.matcher = (matcher != null) ? matcher : new ToscaDefaultMatcher();
-	}
+    public ToscaComparatorProxy(IToscaMatcher matcher) {
+        this.matcher = (matcher != null) ? matcher : new ToscaDefaultMatcher();
+    }
 
-	public int compareTypeCompatible(ToscaNode left, ToscaNode right) {
-		return matcher.isCompatible(left, right) ? 0 : -1;
-	}
+    public int compareTypeCompatible(ToscaNode left, ToscaNode right) {
+        return matcher.isCompatible(left, right) ? 0 : -1;
+    }
 
-	public int compareTypeCompatible(ToscaEdge left, ToscaEdge right) {
-		return matcher.isCompatible(left, right) ? 0 : -1;
-	}
+    public int compareTypeCompatible(ToscaEdge left, ToscaEdge right) {
+        return matcher.isCompatible(left, right) ? 0 : -1;
+    }
 }
