+++ org.eclipse.winery.topologygraph/src/main/java/org/eclipse/winery/topologygraph/matching/ToscaDefaultMatcher.java
@@ -13,20 +13,20 @@
  ********************************************************************************/
 package org.eclipse.winery.topologygraph.matching;
 
-import org.eclipse.winery.topologygraph.matching.model.ToscaEdge;
-import org.eclipse.winery.topologygraph.matching.model.ToscaNode;
+import org.eclipse.winery.topologygraph.model.ToscaEdge;
+import org.eclipse.winery.topologygraph.model.ToscaNode;
 
 import org.apache.commons.lang3.StringUtils;
 
 public class ToscaDefaultMatcher implements IToscaMatcher {
 
-	@Override
-	public boolean isCompatible(ToscaNode left, ToscaNode right) {
-		return StringUtils.equals(left.getNodeTemplate().getName(), right.getNodeTemplate().getName());
-	}
+    @Override
+    public boolean isCompatible(ToscaNode left, ToscaNode right) {
+        return StringUtils.equals(left.getNodeTemplate().getName(), right.getNodeTemplate().getName());
+    }
 
-	@Override
-	public boolean isCompatible(ToscaEdge left, ToscaEdge right) {
-		return StringUtils.equals(left.getTemplate().getName(), right.getTemplate().getName());
-	}
+    @Override
+    public boolean isCompatible(ToscaEdge left, ToscaEdge right) {
+        return StringUtils.equals(left.getTemplate().getName(), right.getTemplate().getName());
+    }
 }
