+++ org.eclipse.winery.topologygraph/src/main/java/org/eclipse/winery/topologygraph/matching/ToscaIsomorphismMatcher.java
@@ -16,22 +16,22 @@ package org.eclipse.winery.topologygraph.matching;
 import java.util.Comparator;
 import java.util.Iterator;
 
-import org.eclipse.winery.topologygraph.matching.model.ToscaEdge;
-import org.eclipse.winery.topologygraph.matching.model.ToscaGraph;
-import org.eclipse.winery.topologygraph.matching.model.ToscaNode;
+import org.eclipse.winery.topologygraph.model.ToscaEdge;
+import org.eclipse.winery.topologygraph.model.ToscaGraph;
+import org.eclipse.winery.topologygraph.model.ToscaNode;
 
 import org.jgrapht.GraphMapping;
 import org.jgrapht.alg.isomorphism.VF2SubgraphIsomorphismInspector;
 
 public class ToscaIsomorphismMatcher {
 
-	public Iterator<GraphMapping<ToscaNode, ToscaEdge>> findMatches(ToscaGraph queryGraph, ToscaGraph toSearchIn, IToscaMatcher comp) {
+    public Iterator<GraphMapping<ToscaNode, ToscaEdge>> findMatches(ToscaGraph queryGraph, ToscaGraph toSearchIn, IToscaMatcher comp) {
 
-		// NEVER CHANGE THE ORDERING OF THE PARAMETERS. THE SWITCH (RIGHT, LEFT) IS DELIBERATE!
-		Comparator<ToscaNode> vertexComparator = (ToscaNode left, ToscaNode right) -> new ToscaComparatorProxy(comp).compareTypeCompatible(right, left);
-		Comparator<ToscaEdge> edgeComparator = (ToscaEdge left, ToscaEdge right) -> new ToscaComparatorProxy(comp).compareTypeCompatible(right, left);
+        // NEVER CHANGE THE ORDERING OF THE PARAMETERS. THE SWITCH (RIGHT, LEFT) IS DELIBERATE!
+        Comparator<ToscaNode> vertexComparator = (ToscaNode left, ToscaNode right) -> new ToscaComparatorProxy(comp).compareTypeCompatible(right, left);
+        Comparator<ToscaEdge> edgeComparator = (ToscaEdge left, ToscaEdge right) -> new ToscaComparatorProxy(comp).compareTypeCompatible(right, left);
 
-		VF2SubgraphIsomorphismInspector<ToscaNode, ToscaEdge> inspector = new VF2SubgraphIsomorphismInspector<>(toSearchIn, queryGraph, vertexComparator, edgeComparator);
-		return inspector.getMappings();
-	}
+        VF2SubgraphIsomorphismInspector<ToscaNode, ToscaEdge> inspector = new VF2SubgraphIsomorphismInspector<>(toSearchIn, queryGraph, vertexComparator, edgeComparator);
+        return inspector.getMappings();
+    }
 }
